requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
django==5.2.1
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.7.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.28.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
packaging==25.0
    # via gunicorn
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via django
uritemplate==4.1.1
    # via drf-spectacular
